sprint18_feat: добавлены ендпойнты начало

+++ main-service/src/main/java/ru/practicum/mainservice/event/model/dto/AdminUpdateEventRequest.java
@@ -27,7 +27,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class AdminUpdateEventRequest {
     private String annotation;
-    private Integer category;
+    private Long category;
     private String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonSerialize(using = LocalDateTimeSerializer.class)
